Temel Operasyonlar

+++ Entities/Concrete/Game.cs
@@ -13,6 +13,6 @@ namespace Entities.Concrete
         public int GameDeveloperId { get; set; }
         public int CategoryId { get; set; }
         public string GameName { get; set; }
-        public double GamePrice { get; set; }
+        public decimal GamePrice { get; set; }
     }
 }
